+++ src/services/home.py
@@ -4,7 +4,7 @@ from src.controllers.authentication import Authentication
 from src.controllers.user import User
 from src.common.flags import Flag
 from src.common.roles import Role
-from src.models import database
+from src.utils import database
 from src.services import blogger
 from src.loggers.general_logger import GeneralLogger
 from src.common import filepaths
